back-end/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
Flask-Session
redis
flask-cors
flask-Migrate
Flask-RESTful
coverage
requests
flask-mail
flasgger
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
Flask-OAuthlib